refactor: 이전 web + docker 관련 파일 제거
