--- docker/docker-compose.yml
@@ -1,21 +1,0 @@
-version: "3.9"
-services:
-  db:
-    image: mysql:8.0.28
-    platform: linux/x86_64
-    restart: always
-    ports:
-      - "3306:3306"
-    environment:
-      MYSQL_ROOT_PASSWORD: root
-      MYSQL_DATABASE: chess
-      MYSQL_USER: user
-      MYSQL_PASSWORD: password
-      TZ: Asia/Seoul
-    volumes:
-      - ./db/mysql/data:/var/lib/mysql
-      - ./db/mysql/config:/etc/mysql/conf.d
-      - ./db/mysql/init:/docker-entrypoint-initdb.d
-
-
-
