+++ src/test/java/chess/dao/jdbctemplate/JdbcTemplateTurnDaoTest.java
@@ -1,4 +1,4 @@
-package chess.dao;
+package chess.dao.jdbctemplate;
 
 import static org.assertj.core.api.Assertions.assertThat;
